Stop Ninja from constantly rebuilding
As explained in the comment, Ninja gets confused by the fact that the
subdirectory is added in `darling/src/libsystem_kernel` but it's
actually located in the root of the repo. The system resolves relative
paths using the real path, but Ninja resolves it according to the CMake
subdirectory path.

Ninja was constantly searching for
`darling/src/libsystem_kernel/osfmk/mach_eventlink_types.h` instead of
`osfmk/mach_eventlink_types.h` and therefore constantly thought that the
source that depended on this was out-of-date.