Use Darling reserved key definitions for TLS keys
Our previous choices for per-thread-WD keys were conflicting with the
keys reserved for the Swift runtime and causing crashes for Swift apps.

Instead, let's use some keys from the newly-reserved Darling range for
pthread keys (200-209). This range seems to be otherwise unused, so it
should be fine for us to use it.