Rework "stat" code to better support ARM64
Unlike i386/x86_64, ARM64 doesn't contain a `stat64` struct. On top of that, ARM64's `stat` is actually the same as `stat64`. As a result, some of the stat code will need to be converted to handle this scenario.

For more details, I recommend looking at the `[xnu]/bsd/sys/stat.h` file.